requirements.txt
Flask>=3.0.0
openai>=1.58.0
python-dotenv>=1.0.0 